Cleaned up and made updates.
Added “bttest” & “hciconfig” xbin files. Both these files had not worked previously but have since changing permissions for and adding “hcitool” The difference in performance in comparison to both using the same files from other variants or the files built from Source. In testing with the files included was able to Transfer large files such as Rom update.zips in less then half the time then with the above comparisons. Also added “hlr_auc_gw” into bin after reading about its relation to the hci files used. This is otherwise an added file and does not replace any files currently built or added.

Added “iwpriv” From Sharp 2.3.5 again after reading about its relation to the hostapd files needed and originating from the sane place or origin. This is otherwise an added file and does not replace any files currently built or added.

Cleaned up and did a bit more labeling of from what variant each file originates. Again would like to re-organize the entire structure, Also cleaned out leftover files either left in the Proprietary folder by either myself or past developers.

Moved  “liboncrpc.so” back into the Sharp 2.3.5 list. Had tested with the Cherry file for a bit based on a few media calls but ultimately see it more in relation to the other files we add from Sharp 2.3.5. Still on the fence about “libmmgsdilib.so” included previously for the same reasons as “liboncrpc.so” but until I learn more left it alone as there does not seem to be a noticeable difference. Lib “libcm.so” is still needed from either the Cherry, WellcoM or Spice Variants as it is the only file outside of those used for data from Stock-Triumph that does not work from Sharp 2.3.5. At this point I would not rule out a performance issue as I am finding the Sharp 2.3.5 files sometimes require different permissions then the other variants but with that said this only seems to holds true so far with files from the bin directory.